daemon/requirements.txt
# Lyra Discord Daemon Dependencies

# Discord connection
discord.py>=2.3.0

# Environment variable loading
python-dotenv>=1.0.0
